code_backup/Internal_network_share/requirements.txt
Flask==3.0.3
waitress==2.1.2
python-dotenv==1.0.1
